Switch to cargo workspace
- Add .editorconfig
- Amend spacing in day-01
- Fix path issue with `day-01/INPUT`
- Add travis config
- Update readme
