+++ .travis.yml
@@ -1,0 +1,15 @@
+language: rust
+
+script:
+  - cargo build --verbose --all
+  - cargo test --verbose --all
+
+rust:
+  - stable
+  - beta
+  - nightly
+
+matrix:
+  allow_failures:
+    - rust: nightly
+  fast_finish: true
