--- 2018/README.md
@@ -1,5 +1,0 @@
-# 2018
-
-Learning Rust from scratch! Pull requests are welcome. Because I'm learning, please feel free to correct style mistakes, share idiomatic patterns, etc.
-
-Thanks!
