--- 2018/day-01/README.md
@@ -1,13 +1,0 @@
-# Day 1
-
-## Running
-
-```
-cargo run
-```
-
-## Tests
-
-```
-cargo test
-```
