+++ README.md
@@ -1,3 +1,17 @@
-# Advent of Code
+# 2018
 
-Repository containing my AoC attempts.
+Learning Rust from scratch! Pull requests are welcome. Because I'm learning, please feel free to correct style mistakes, share idiomatic patterns, etc.
+
+Thanks!
+
+## Running
+
+```
+cargo run -p <day-xx> --release
+```
+
+## Tests
+
+```
+cargo test
+```
